IDEA-73175 While editing, don't mark file dirty in another project based on the same root
PROBLEM DESCRIPTION
On document change, VcsFileStatusProvider checks its file status and in the case of null, marks it dirty. Once the status if requested (FileStatusManager.getStatus), it is calculated and saved in the map.
In other project [based on the same VCS root] the file is marked dirty on each document change, because its status is never queried.

SOLUTION
When the status is null, save it in the map with a fake FileStatusNull.INSTANCE status.
Once the file is queried, correct status will be calculated. Otherwise (for the unrelated project) it just will remain in the map, and the file won't be marked dirty each time.

PROBLEM 2
On each ChangeListManagerImpl.updateImmediately the file statuses map is cleared, eliminating the solution above. That's because equals is not overridden in IgnoredFilesCompositeHolder, and so the check FileCompositeHolder#equals in updateImmediately always fails (because it uses Object#equals).

SOLUTION 2
Implement IgnoredFilesCompositeHolder#equals comparing the only significant field - the map of ignored files.

[Reviewed by Max]