Fix Git push to alternative remote
IDEA-78244:
GitPushDialog#getTrackedCurrentBranchName -> getRemoteAndTrackedBranchForCurrentBranch
When pre-filling values in the "push to manual branch panel", fill not only the tracked branch name, but select the tracked remote as well. In case of error or no tracked branch, use "origin" as the default remote. GitPushDialog.getNameWithoutRemote: use the discovered tracked remote name, not the pre-selected.
Don't update the remote and branch fields if checkbox is selected: user is making his choice.

IDEA-77985:
GitPushDialog#pushSpecsForCurrentOrEnteredBranches : use the selected remote.

IDEA-78245:
Increase text field size to 20 for that unlikely (after above fixes) case when the full branch name would be written, or when branch name is long.