CherryPicker: commit more synchronously
Just calling AbstractVcsHelper#commit makes commit be executed asynchronously, and cherry-pick continues while commit has not yet be performed.
That way, instead of using the CommitHelper, let supply own CommitExecutor that calls GitCheckinEnvironment#commit.
CommitExecutors are called synchronously, under a modal progress, that way the AWT thread would wait for the commit to complete, and CherryPicker is waiting for the AWT.