IDEA-85560 Files are not refreshed and Pull Info is not shown after Git pull
Problem:
since pull is run in the background, GitRepositoryAction starts the final tasks (VFS refresh, delayed tasks execution) before the task completes.

Solution:
Introduce the flag-method executeFinalTasksSynchronously, and return false from the actions which are asynchronous and should call the tasks themselves.
These actions are pull and fetch (final tasks are not needed for the latter).
Execute delayed tasks of GitMergeUtil in AWT later, because now they are called on a pooled thread on certain actions.