[git] Cherry pick: use the common commit dialog.
Instead of using a custom commit executor & custom commit session, use the common commit dialog.
The problem is that the commit dialog "onOK"-procedure is asynchronous.
To fix this provide a CommitResultHandler that is called on successful or failed commit, and wait (in the GitCherryPicker) until it reports that commit has completed.
Don't show notification or report error during commit if the CommitResultHandler is set.

This fixes IDEA-84825, IDEA-88236 and is necessary for IDEA-87476.