[git] cucumber tests: move from groovy to java, introduce virtual commits, understand commit details syntax
Virtual commits are to be able to specify commit hashes in feature files.
Since it's not possible to can't assign a hash to a commit, remember the actual hash after committing and link them together.

Make it able to specify commit details in the feature file similar to the way they are written in the log, parse it using CommitDetails, and perform the commit given these specifications.