[github] IDEA-100618 Support self-signed SSL certificates
Catch the ValidatorException which is thrown when the server's SSL
certificate is not trusted (self-signed, for instance) and show a
message to the user proposing to proceed anyway (as it is done in
browsers).
If user agrees to proceed, remember the host and add it to "trusted"
hosts; next time it is queried don't ask the user and just proceed
without certificate verification.

Implement all this logic in the GithubSslSupport using the
EasySSLProtocolSocketFactory to perform an unchecked connection.

Store trusted hosts in GithubSettings. It uses some old-style
PersistentComponent creation, just following the existing style for now.