[vcs] Compact and move the Commit Dialog "Summary" section
IDEA-90202 To save place at the right panel of the commit dialog
compact the "Summary" section:
* CommitLegendPanel: instead of a complex 3x4 grid use something as
light as "New: 1 of 20  Deleted: 3 of 30  Modified: 4 of 4"
  Don't say "1 of 1" if all changes are included.
* Move the section beyond the changes browser,
above the commit message area.
* Adjust the default proportion of the splitter between changes
browser and commit message to let it look better on small window sizes.

After the summary section was compacted, it started to look really
empty at the right panel of the Revert dialog and the Apply Patch
dialog.
To fix this in the RollbackChangesDialog:
* Move the changelist selector to the changes browser toolbar (as was
 already done for the commit dialog).
* Move the summary & "delete locally created files" below the changes
 browser.

To fix this in the ApplyPatchDifferentiatedDialog move "summary"
below the changes browser as well. Also, remove the second
GridBagLayouted wrapper panel, and simplify grid bag constraints usage.